tests: move unit tests into src
